+++ R/helper_functions.R
@@ -99,12 +99,14 @@ replicate64 <- function(n, expr, simplify = "array") {
            simplify = simplify)
 }
 
+#nocov start
 #' Produce a legend for continuous gradient scales
 #' 
 #' Prints an annotated vertical bar coloured according to a continuous palette.
 #' 
 #' This function is now deprecated; it has been superseded by the more capable
 #' [`PlotTools::SpectrumLegend()`] and will be removed in a future release.
+# Deprecation notice added in TreeTools 1.9.2 (2023-04-25)
 #' 
 #' @param x0,y0,x1,y1 Coordinates of the bottom-left and top-right end of the
 #' bar.
@@ -184,3 +186,4 @@ SpectrumLegend <- function(x0 = 0.05, y0 = 0.05,
          ...)
   }
 }
+#nocov end