+++ R/tree_rearrangement.R
@@ -556,6 +556,7 @@ MakeTreeBinary.phylo <- function(tree) {
   }
   tree[["edge"]] <- edge
   tree[["Nnode"]] <- nTip - 1L
+  attr(tree, "order") <- NULL
   tree
 }