+++ inst/include/TreeTools/keep_tip.h
@@ -10,9 +10,9 @@
 
 #define RCOUT Rcpp::Rcout
 
-#define RETAIN 9000
+constexpr int RETAIN = 9000;
 
-#define GET_NEW_NO(n)  if (!new_no[n]) new_no[n] = ++next_no
+#define GET_NEW_NO(n) if (!new_no[n]) new_no[n] = ++next_no
 
 #define SKIP_EDGE                                              \
 GET_NEW_NO(parent);                                            \