+++ src/descendant_edges.cpp
@@ -2,7 +2,7 @@
 #include <stdexcept> /* for errors */
 using namespace Rcpp;
 
-#define R_TO_C 1
+constexpr int R_TO_C = 1;
 
 // [[Rcpp::export]]
 LogicalMatrix descendant_edges(