+++ src/tips_in_splits.cpp
@@ -4,7 +4,7 @@
 #include "../inst/include/TreeTools.h"
 using namespace Rcpp;
 
-const uint_fast32_t powers_of_two[32] = {
+constexpr uint_fast32_t powers_of_two[32] = {
   0x1, 0x2, 0x4, 0x8,
   0x10, 0x20, 0x40, 0x80,
   0x100, 0x200, 0x400, 0x800,