+++ tests/testthat/test-helper_functions.R
@@ -9,29 +9,3 @@ test_that("UnshiftTree() works", {
   attr(expectation, "tip.label") <- NULL
   expect_equal(expectation, UnshiftTree(t1, as.phylo(2, 8)))
 })
-
-
-test_that("SpectrumLegend()", {
-  skip_if(packageVersion("graphics") < "4.1")
-  skip_if(packageVersion("vdiffr") < "1.0")
-  vdiffr::expect_doppelganger("SpectrumLegend", function() {
-    
-    # Set up blank plot
-    plot(0:1, 10:11, asp = 1, type = "n", frame.plot = FALSE,
-         xlab = "x", ylab = "y")
-    expect_warning(
-      SpectrumLegend(legend = c("Bottom", "Middle", "Top"),
-                     lwd = 5,
-                     palette = rev(hcl.colors(256L, "RdYlBu")),
-                     text.col = c("blue", "brown", "red"),
-                     title = "Blue title", title.font = 3, title.cex = 0.5),
-    "'SpectrumLegend' is deprecated.")
-    expect_warning(
-      SpectrumLegend(0.4, 10.6, 0.2, 11, abs = TRUE,
-                     legend = seq(0, 10, by = 2), palette = 0:10,
-                     lty = "dotted", pos = 2,
-                     lmitre = 2, lend = "round", ljoin = "round",
-                     title = "Multi-\nline\ntitle", xpd = NA),
-    "'SpectrumLegend' is deprecated.")
-  })
-})